refactor(Sections): extract class which is getting stations ids
- 역 목록을 조회하는 책임을 클래스 분리
