+++ src/main/java/nextstep/subway/line/domain/sections/Sections.java
@@ -43,21 +43,6 @@ public class Sections {
         this.sections.add(section);
     }
 
-    public List<Long> getStationIdsOrderBySection() {
-        List<Long> stationIds = new ArrayList<>();
-        Section endUpSection = findEndUpSection();
-        stationIds.add(endUpSection.getUpStationId());
-        stationIds.add(endUpSection.getDownStationId());
-
-        Section nextSection = this.findNextSection(endUpSection);
-        while (nextSection != null) {
-            stationIds.add(nextSection.getDownStationId());
-            nextSection = this.findNextSection(nextSection);
-        }
-
-        return stationIds;
-    }
-
     public Section findTargetSection(final Section newSection) {
         Section targetSection = findSameWithUpStation(newSection);
         if (targetSection == null) {
